move exec back
