+++ app/src/main/kotlin/no/nav/arbeidsgiver/notifikasjon/Main.kt
@@ -1,11 +1,6 @@
 package no.nav.arbeidsgiver.notifikasjon
 
-import no.nav.arbeidsgiver.notifikasjon.bruker.Bruker
-import no.nav.arbeidsgiver.notifikasjon.ekstern_varsling.EksternVarsling
 import no.nav.arbeidsgiver.notifikasjon.infrastruktur.logger
-import no.nav.arbeidsgiver.notifikasjon.kafka_reaper.KafkaReaper
-import no.nav.arbeidsgiver.notifikasjon.produsent.Produsent
-import no.nav.arbeidsgiver.notifikasjon.statistikk.Statistikk
 
 @Suppress("unused") /* Entry point for jar. */
 private object Main {
