+++ app/src/test/kotlin/no/nav/arbeidsgiver/notifikasjon/bruker_api/BrukerModelIdempotensTests.kt
@@ -2,7 +2,7 @@ package no.nav.arbeidsgiver.notifikasjon.bruker_api
 
 import io.kotest.core.datatest.forAll
 import io.kotest.core.spec.style.DescribeSpec
-import no.nav.arbeidsgiver.notifikasjon.bruker.Bruker
+import no.nav.arbeidsgiver.notifikasjon.Bruker
 import no.nav.arbeidsgiver.notifikasjon.HendelseModel.Hendelse
 import no.nav.arbeidsgiver.notifikasjon.bruker.BrukerRepositoryImpl
 import no.nav.arbeidsgiver.notifikasjon.util.EksempelHendelse
