+++ app/src/test/kotlin/no/nav/arbeidsgiver/notifikasjon/executable/bruker_api/LocalMainBrukerAPI.kt
@@ -1,7 +1,7 @@
 package no.nav.arbeidsgiver.notifikasjon.executable.bruker_api
 
 import db.migration.OS
-import no.nav.arbeidsgiver.notifikasjon.bruker.Bruker
+import no.nav.arbeidsgiver.notifikasjon.Bruker
 import no.nav.arbeidsgiver.notifikasjon.bruker.BrukerModel
 import no.nav.arbeidsgiver.notifikasjon.bruker.BrukerModel.Tilganger
 import no.nav.arbeidsgiver.notifikasjon.infrastruktur.http.HttpAuthProviders
