+++ app/src/test/kotlin/no/nav/arbeidsgiver/notifikasjon/executable/ekstern_varsling/LocalMainEksternVarsling.kt
@@ -1,7 +1,7 @@
 package no.nav.arbeidsgiver.notifikasjon.executable.ekstern_varsling
 
 import db.migration.OS
-import no.nav.arbeidsgiver.notifikasjon.ekstern_varsling.EksternVarsling
+import no.nav.arbeidsgiver.notifikasjon.EksternVarsling
 
 fun main(@Suppress("UNUSED_PARAMETER") args: Array<String>) {
     OS.setupLocal()
