+++ app/src/test/kotlin/no/nav/arbeidsgiver/notifikasjon/executable/kafka_reaper/LocalMainKafkaReaper.kt
@@ -1,7 +1,7 @@
 package no.nav.arbeidsgiver.notifikasjon.executable.kafka_reaper
 
 import db.migration.OS
-import no.nav.arbeidsgiver.notifikasjon.kafka_reaper.KafkaReaper
+import no.nav.arbeidsgiver.notifikasjon.KafkaReaper
 
 /* Bruker API */
 fun main(@Suppress("UNUSED_PARAMETER") args: Array<String>) {
