+++ app/src/test/kotlin/no/nav/arbeidsgiver/notifikasjon/executable/statistikk/LocalMainStatistikk.kt
@@ -1,7 +1,7 @@
 package no.nav.arbeidsgiver.notifikasjon.executable.statistikk
 
 import db.migration.OS
-import no.nav.arbeidsgiver.notifikasjon.statistikk.Statistikk
+import no.nav.arbeidsgiver.notifikasjon.Statistikk
 
 /* Statistikk */
 fun main(@Suppress("UNUSED_PARAMETER") args: Array<String>) {
