+++ app/src/test/kotlin/no/nav/arbeidsgiver/notifikasjon/kafka_reaper/KafkaReaperModelTest.kt
@@ -3,6 +3,7 @@ package no.nav.arbeidsgiver.notifikasjon.kafka_reaper
 import io.kotest.core.datatest.forAll
 import io.kotest.core.spec.style.DescribeSpec
 import no.nav.arbeidsgiver.notifikasjon.HendelseModel.Hendelse
+import no.nav.arbeidsgiver.notifikasjon.KafkaReaper
 import no.nav.arbeidsgiver.notifikasjon.util.EksempelHendelse
 import no.nav.arbeidsgiver.notifikasjon.util.testDatabase
