+++ app/src/test/kotlin/no/nav/arbeidsgiver/notifikasjon/produsent/api/IdempotensOppførselForProdusentApiTests.kt
@@ -4,12 +4,8 @@ import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.mockk
 import no.nav.arbeidsgiver.notifikasjon.HendelseModel.AltinnMottaker
-import no.nav.arbeidsgiver.notifikasjon.produsent.Produsent
-import no.nav.arbeidsgiver.notifikasjon.produsent.api.ProdusentAPI
+import no.nav.arbeidsgiver.notifikasjon.Produsent
 import no.nav.arbeidsgiver.notifikasjon.produsent.ProdusentRepositoryImpl
-import no.nav.arbeidsgiver.notifikasjon.produsent.api.Error
-import no.nav.arbeidsgiver.notifikasjon.produsent.api.MutationNyBeskjed
-import no.nav.arbeidsgiver.notifikasjon.produsent.api.MutationNyOppgave
 import no.nav.arbeidsgiver.notifikasjon.util.*
 import java.util.*
